Consolidated MPFA-O mesh source and shook TDy out of it.
