+++ src/mpfao/tdympfao.c
@@ -1,7 +1,6 @@
 #include <tdytimers.h>
 #include <private/tdycoreimpl.h>
 #include <private/tdymeshimpl.h>
-#include <private/tdymeshutilsimpl.h>
 #include <private/tdyutils.h>
 #include <private/tdymemoryimpl.h>
 #include <private/tdympfaoutilsimpl.h>
